requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
Pillow